Working on

+++ foody_restaurant/src/main/java/com/example/foodyrestaurant/MenuEditItem.java
@@ -19,6 +19,7 @@ import android.support.v7.widget.RecyclerView;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.Log;
+import android.view.Menu;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
@@ -80,6 +81,10 @@ public class MenuEditItem extends AppCompatActivity {
         if (dialogDism != null){
             dialogDism.dismiss();
         }
+
+        fileTmp = new File(storageDir, JSON_COPY);
+        String json = jsonHandler.toJSON(cards);
+        jsonHandler.saveStringToFile(json, fileTmp);
     }
 
     @Override
@@ -87,35 +92,43 @@ public class MenuEditItem extends AppCompatActivity {
         super.onRestoreInstanceState(savedInstanceState);
         String dialogPrec = savedInstanceState.getString("dialog");
         posToChange = savedInstanceState.getInt("posToChange", 0);
+        fileTmp = new File(storageDir, JSON_COPY);
         cards = jsonHandler.getCards(fileTmp);
-        dishes = getDishes();
+        placeholderPath = savedInstanceState.getString("placeholderPath");
+        dishes = getDishes(JSON_COPY);
         unchanged = savedInstanceState.getBoolean("unchanged");
         if (dialogPrec != null && dialogPrec.compareTo("ok") != 0) {
             if (dialogPrec.compareTo("back") == 0) {
                 restoreBack();
             }
         }
+        for(int i = 0; i < dishes.size();i ++){
+            Log.d("TITLECHECK",dishes.get(i).toString());
+        }
+        recyclerAdapter.notifyDataSetChanged();
     }
 
     @Override
     protected void onSaveInstanceState(Bundle outState) {
         super.onSaveInstanceState(outState);
-        fileTmp = new File(storageDir, JSON_COPY);
         outState.putBoolean("unchanged", unchanged);
         outState.putString("dialog", dialogCode);
         outState.putInt("posToChange", posToChange);
+        outState.putString("placeholderPath", placeholderPath);
 
         for (int i = 0; i < cards.size(); i++){
             if(cards.get(i).getTitle().equals(className)){
                 cards.get(i).setDishes(dishes);
             }
         }
+        fileTmp = new File(storageDir, JSON_COPY);
         String json = jsonHandler.toJSON(cards);
         jsonHandler.saveStringToFile(json, fileTmp);
     }
 
     private void init(){
         storageImageDir = getExternalFilesDir(Environment.DIRECTORY_PICTURES);
+
         file = new File(storageDir, JSON_PATH);
         jsonHandler = new JsonHandler();
         final RecyclerView recyclerMenu = findViewById(R.id.menu_items);
@@ -128,7 +141,11 @@ public class MenuEditItem extends AppCompatActivity {
         TextView title = findViewById(R.id.textView);
         className = getIntent().getExtras().getString("MainName");
         title.setText(getResources().getString(R.string.edit, className));
-        dishes = getDishes();
+        File temp = new File(storageDir,JSON_COPY);
+        if(temp.exists())
+            dishes = getDishes(JSON_COPY);
+        else
+            dishes = getDishes(JSON_PATH);
         recyclerAdapter = new RVAdapterEditItem(dishes, this);
         recyclerMenu.setAdapter(recyclerAdapter);
         fabDishes = findViewById(R.id.fabDishes);
@@ -156,23 +173,12 @@ public class MenuEditItem extends AppCompatActivity {
         ArrayList<Card> oldCards = jsonHandler.getCards(oldFile);
         ArrayList<Dish> oldDishes = new ArrayList<>();
 
-        Log.d("TITLECHECK","OldCards");
-        for(int i = 0; i < oldCards.size(); i++){
-            oldCards.get(i).print();
-        }
-
-        Log.d("TITLECHECK","Cards");
-        for(int i = 0; i < cards.size(); i++){
-            cards.get(i).print();
-        }
-
-
-
         for (int i = 0; i < cards.size(); i++){
             if(cards.get(i).getTitle().equals(className)){
                 oldDishes = oldCards.get(i).getDishes();
             }
         }
+
 
         boolean stillExists;
         for(int i = 0; i < oldDishes.size(); i++){
@@ -182,7 +188,7 @@ public class MenuEditItem extends AppCompatActivity {
                     stillExists = true;
                 }
             }
-            if(stillExists == false){
+            if(stillExists == false && oldDishes.get(i).getImage() != null){
                 File toDelete = new File(oldDishes.get(i).getImage().getPath());
                 if(toDelete.exists())
                     toDelete.delete();
@@ -191,7 +197,6 @@ public class MenuEditItem extends AppCompatActivity {
 
 
         String json = jsonHandler.toJSON(cards);
-        Log.d("TITLECHECK",json);
         File file = new File(storageDir, JSON_REAL);
         File fileTMP = new File(storageDir, JSON_PATH);
         File fileItem = new File(storageDir, JSON_COPY);
@@ -247,10 +252,10 @@ public class MenuEditItem extends AppCompatActivity {
         recyclerAdapter.notifyItemRangeChanged(position, dishes.size());
     }
 
-    private ArrayList<Dish> getDishes(){
+    private ArrayList<Dish> getDishes(String fileName){
 
         ArrayList<Dish> dishes = new ArrayList<>();
-        File plc = new File(storageDir, JSON_PATH);
+        File plc = new File(storageDir, fileName);
         cards = jsonHandler.getCards(plc);
 
         for (int i = 0; i < cards.size(); i++){
@@ -542,5 +547,11 @@ public class MenuEditItem extends AppCompatActivity {
             return Uri.fromFile(fileToSave);
         }
         return null;
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        recyclerAdapter.notifyDataSetChanged();
     }
 }
